update (still develop)

+++ src/pages/product/create.jsx
@@ -2,6 +2,7 @@ import React, { useContext, useEffect, useRef, useState } from 'react'
 import { useNavigate } from 'react-router-dom'
 import {LazyLoadImage} from "react-lazy-load-image-component"
 import convertPrice from '../../../utils/price'
+import alert from "../../../utils/alert"
 import swal from "sweetalert2"
 import axios from "axios"
 import "../../style/create.css"
@@ -25,22 +26,27 @@ const Create = () => {
   }
 
   const createProduct = async () => {
-    try {
-      let formData = new FormData()
-      formData.append('ctg', ctg);
-      formData.append('img', image);
-      formData.append('desc', desc);
-      formData.append('file', file);
-      formData.append('title', title);
-      formData.append('price', price);
-      const response = await axios.post(`${import.meta.env.VITE_API}/product`,formData, {
-        headers: {"Content-Type": 'multipart/form-data'}
-      })
-      localStorage.clear()
-      swal.fire({icon:'success',text:response.data,showConfirmButton:false,timer:2500})
-      .then((res) => res.dismiss && location.reload())
-    } catch (error) {
-      if (error.response) {swal.fire({icon: 'error', text:error.response.data,showConfirmButton:false})}
+    
+    if (file && title && image && desc && price && ctg) {
+      try {
+        let formData = new FormData()
+        formData.append('ctg', ctg);
+        formData.append('img', image);
+        formData.append('desc', desc);
+        formData.append('file', file);
+        formData.append('title', title);
+        formData.append('price', price);
+        const response = await axios.post(`${import.meta.env.VITE_API}/product`,formData, {
+          headers: {"Content-Type": 'multipart/form-data'}
+        })
+        localStorage.clear()
+        swal.fire({icon:'success',text:response.data,showConfirmButton:false,timer:2500})
+        .then((res) => res.dismiss && location.reload())
+      } catch (error) {
+        if (error.response) { alert(error.response.data) }
+      }
+    } else {
+      alert("please complete the form data!")
     }
   }
 
@@ -65,30 +71,33 @@ const Create = () => {
             <div>Price :</div>
             <input className='productinput' value={price} type="text" placeholder='ex: 350000' onChange={(e) => setPrice(e.target.value)}/>
           </div>
-          <div className='wrap-file'>
-            <div>
-              <div>Preview : </div>
-              <div className='prevfile' onClick={() => imgref.current.click()}>
-                <div className={(image) ? 'fa-solid fa-check fa-xl' : 'fa-solid fa-image fa-xl'} style={{color: 'var(--background)'}}/>
-                <input type="file" onChange={(e) => setImage(e.target.files[0])} ref={imgref} style={{display:'none'}}/>
-              </div>
-            </div>
-            <div>
-              <div>File : </div>
-              <div className='prevfile' onClick={() => fileref.current.click()}>
-                <div className={(file) ? 'fa-solid fa-check fa-xl' : 'fa-solid fa-file fa-xl'} style={{color: 'var(--background)'}}/>
-                <input type="file" onChange={(e) => setFile(e.target.files[0])} ref={fileref} style={{display:'none'}}/>
-              </div>
-            </div>
-            <div>
+          <div>
               <div>Category :</div>
-              <select value={ctg} onChange={(e) => setCtg(e.target.value)}>
+              <select style={{width: '100%'}} value={ctg} onChange={(e) => setCtg(e.target.value)}>
                 <option value=""></option>
                 <option value="web">Web</option>
                 <option value="vector">Vector</option>
                 <option value="video">Video</option>
               </select>
             </div>
+          <div className='wrap-file'>
+            <div>
+              <div>Preview : </div>
+              <div className='prevfile' onClick={() => imgref.current.click()}>
+                <div className={(image) ? 'fa-solid fa-check fa-xl' : 'fa-solid fa-image fa-xl'} style={{color: 'var(--background)', fontSize: '2rem'}}/>
+                <div style={{ color: 'var(--background)', fontSize: '0.9rem' }}>{'(JPEG, JPG, PNG, MP4)'}</div>
+                <input type="file" onChange={(e) => setImage(e.target.files[0])} ref={imgref} style={{display:'none'}}/>
+              </div>
+            </div>
+            <div>
+              <div>File : </div>
+              <div className='prevfile' onClick={() => fileref.current.click()}>
+                <div className={(file) ? 'fa-solid fa-check fa-xl' : 'fa-solid fa-file fa-xl'} style={{color: 'var(--background)', fontSize: '2rem'}}/>
+                <input type="file" onChange={(e) => setFile(e.target.files[0])} ref={fileref} style={{display:'none'}}/>
+                <div style={{ color: 'var(--background)', fontSize: '0.9rem' }}>{'(ZIP, RAR)'}</div>
+              </div>
+            </div>
+            <div className='button-max' onClick={() => createProduct()} style={(file && title && image && desc && price && ctg) ? {backgroundColor: 'var(--yellow)', marginTop: '60px'} : {backgroundColor: '#aaa', marginTop: '60px'}}>Create</div>
             </div>
           </div>
         <div className='prev-form'>
@@ -104,7 +113,6 @@ const Create = () => {
               </div>
             </div>
           </div>
-          <div className='button' onClick={() => createProduct()} style={(file && title && image && desc && price && ctg) ? {width:'130px', marginTop:"30px"} : {display:'none'}}>Create</div>
         </div>
       </div>
     </div>
